Go back to uint32

+++ tests/modules/FixedPriceSignatureMinter.t.sol
@@ -668,7 +668,7 @@ contract FixedPriceSignatureMinterTests is TestConfig {
 
         uint32 expectedStartTime = 123;
         uint32 expectedEndTime = 502370;
-        uint32 expectedPrice = 1234071;
+        uint96 expectedPrice = 1234071;
 
         minter.createEditionMint(
             address(edition),
